refactor(ponder-sdk): move functionality from `ensindexer` to `ponder-sdk`
