--- apps/ensindexer/src/lib/indexing-status/ponder-metadata/index.ts
@@ -1,6 +1,0 @@
-export * from "./block-refs";
-export * from "./chains";
-export * from "./config";
-export * from "./metrics";
-export * from "./rpc";
-export * from "./status";
