+++ packages/ponder-sdk/src/prometheus/prometheus-metrics.ts
@@ -1,6 +1,6 @@
 import parsePrometheusTextFormat, { type PrometheusMetric } from "parse-prometheus-text-format";
 // Ensures local declaration file is available to downstream consumers
-import "./types/parse-prometheus-text-format";
+import "./parse-prometheus-text-format";
 
 interface ParsedPrometheusMetric extends Omit<PrometheusMetric, "metrics"> {
   metrics: Array<{
